feat: 가격 Int > Long 변경

+++ app/src/main/java/com/store_me/storeme/data/MenuData.kt
@@ -3,9 +3,9 @@ package com.store_me.storeme.data
 data class MenuData(
     val name: String,
     val priceType: String,
-    val price: Int?,
-    val minPrice: Int?,
-    val maxPrice: Int?,
+    val price: Long?,
+    val minPrice: Long?,
+    val maxPrice: Long?,
     val image: String?,
     val description: String?,
     val isSignature: Boolean,
